fix: handle popular links
The table `publisher.editions` is currently empty, because the column
`url` isn't nullable, but the query of the Publisher database is
extracting null values in that column.

Since August 2024, there is a new, special-case document that has no
slug, which is "popular links". That document is rendered as part of the
homepage. When it is published to the Publishing API, its Content ID is
hardcoded to `ad7968d0-0339-40b2-80bc-3ea1db8ef1b7`, and its base_path
is `null`. The results of this query are joined to the Publishing API
editions by base_path (or url, whatever), and null can't be joined to
`null`. So, even if we were to allow nulls in the url/base_path column
in BigQuery, we'd have to hardcode a special case to make the join work
for the "popular links" document. It isn't worthwhile, so the document
is omitted here.

https://github.com/alphagov/publisher/blob/a01d44291a85e7809cc0f85273c26fea5a7aed3a/app/models/popular_links_edition.rb#L49

This commit undoes a change that was accidentally included in a previous
commit, which was to substitute an empty string for a null slug. But
this was never deployed. If it had been, then the `publisher.editions`
table wouldn't have been empty, but it would have been misleading,
because the rows of the "popular links" document would have been joined
to the rows of the homepage itself.