refactor: migrate all cloud functions to Cloud Run
Close #749

Permadiffs cause `terraform apply` to redeploy cloud functions
unnecessarily, and it is currently also unbinding service accounts from
IAM permissions, which has caused the overnight batch of
`publishing-api` to fail. It also seems to interfere with the
`google_bigquery_job` that registers the remote function in BigQuery.

Now that Cloud Functions are a wrapper around Cloud Run, there would be
no functional difference if we were to refactor the Cloud Functions as
Cloud Run instances directly. The Dockerfile would make testing easier,
and would be offset by being able to delete the double-handling in
Terraform of creating both a Cloud Run instance and a Cloud Function
wrapper. An existing Cloud Run instance to use as a pattern is
terraform/html-to-text.tf and the corresponding source code
docker/html-to-text.