Fix escaped characters in whitehall data pipeline attachments table
Some of the values we receive in the title field look like `Peer review of  \"some other thing\"` or `string \n string` (for long titles), tripping up our bq load from the csv. This commit replaces any occurrences of `\"` with `` and any occurrences of `\n` with ` `. A slight mismatch in the title should be ok, since we mostly care about matching IDs for the monitoring tool we're using the data pipeline for.