Extract data from mysql dump as CSV for BigQuery
These are scraps of scripts that haven't been documented. The idea is to
extract data directly from the dump file, similarly to pg_restore. The
mysql dump file is a gzipped SQL file. The scripts scan the INSERT
statements character-by-character, keeping track of quotes and escapes,
and emit the same information as a CSV that BigQuery can natively
import. If the SQL dump files change much, then the scripts will break
-- they aren't robustly parsing SQL.