feat: mask_pii() BigQuery custom function
Implements #780

The call to the DLP (Data Loss Prevention) API is done by a Cloud Run
Service. A BigQuery remote function called data_loss_prevention() wraps
that. Then a second BigQuery custom function called mask_pii() wraps the
data_loss_prevention() function.

The mask_pii() function is expected to be used most often. It truncates
strings to the maximum byte length that is supported by the DLP API, and
it hardcodes an inspect_config and deidentify_config that have been used
for a few years for masking PII in feedback. If a different
inspect_config and deidentify_config are required, then the
data_loss_prevention() function can be used directly.