fix: allow BigQuery phone number function access
It isn't documented whether anyone needs any role for a BigQuery
user-defined function to be able to fetch a Javascript library
dependency from a bucket, but my guess is that any user of the function,
or any service account that runs a scheduled query that uses the
function, must have roles/storage.objectViewer. And someone who didn't
have access to that bucket also couldn't run the function.