fix: http-to-bucket handle single JSON object
The http-to-bucket Cloud Run service fetches a JSON response from an
API, and then formats it as newline-delimited JSON (ndjson AKA
JSONlines) for uploading into BigQuery.

Originally this was written and tested with the Smart Survey API, which
returns an array of JSON objects, e.g.

```json
[{"foo": 1, "bar": true},{"foo": 1, "bar": true}]
```

and reformatted it as

```json
{"foo": 1, "bar": true}
{"foo": 1, "bar": true}
```

But the Zendesk API returns a single JSON object, e.g.

```json
{"foo": 1, "bar": true}
```

And the code accidentally formatted this as follows.

```json
[{"foo": 1},{"bar": true}]
```

This pull request handles both cases, by checking whether the API has
returned an array or a singleton object.