Fix URL field after publisher postgres move
We recently made some changes to allow the knowledge graph
to support Mainstream Publisher changes following the apps
move to postgres from Mongo. There was an oversight in the
move that has caused data to not be recorded. This should
fix that by creating a field that was present in Mongo that
isn't in postgres