IA-1746: fix rsync issue caused by previous commit
Changes to the deployment process were made in commit 0b84130f979113f4b86e22e862b4cdbe6e6980f5.
The image deployment worked as expected but the refactored rsync of the repo to GCS had a bug.
This change creates a separate workflow for the rsync process so that it behaves as expected.
The downside is that two approvals will be required for production deployments as the image and rsync deployments are sepaate jobs.